requirements.txt
requests
# pydeck
pandas
streamlit
Pillow
plotly
matplotlib
seaborn
